requirements.txt
chandra-ocr[hf]
torch
mss
Pillow
opencv-python
rapidfuzz
keyboard
requests
beautifulsoup4
selenium
webdriver-manager
